Move httpx code into the internal package
