+++ internal/httpx/httpx_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/apex/log"
-	"github.com/ooni/probe-engine/httpx/httpx"
+	"github.com/ooni/probe-engine/internal/httpx"
 )
 
 func TestGet(t *testing.T) {
